fix host IMA

+++ python/sglang/srt/managers/scheduler.py
@@ -2044,7 +2044,7 @@ class Scheduler(
         )
 
         batch.req_pool_indices = torch.tensor(
-            [r.req_pool_idx for r in reqs], dtype=torch.int32, device=device
+            [r.req_pool_idx for r in reqs], dtype=torch.int64, device=device
         )
         seq_lens = [len(r.origin_input_ids) + len(r.output_ids) - 1 for r in reqs]
         batch.seq_lens = torch.tensor(seq_lens, dtype=torch.int64, device=device)
